# packages/api-client/openapi.json
{
	"openapi": "3.0.0",
	"paths": {
...
		"/api/teams": {
			"get": {
				"operationId": "teamsList",
				"parameters": [],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TeamStateDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "List teams and resolve active team",
				"tags": ["teams"]
			},
			"post": {
				"operationId": "teamsCreate",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateTeamDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TeamDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Create a team",
				"tags": ["teams"]
			}
		},
		"/api/teams/active": {
			"put": {
				"operationId": "teamsSetActive",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SetActiveTeamDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ActiveTeamResponseDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Set active team",
				"tags": ["teams"]
			}
		},
		"/api/teams/{teamId}": {
			"patch": {
				"operationId": "teamsRename",
				"parameters": [
					{
						"name": "teamId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RenameTeamDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TeamDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Rename a team",
				"tags": ["teams"]
			},
			"delete": {
				"operationId": "teamsDelete",
				"parameters": [
					{
						"name": "teamId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OkDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Delete a team",
				"tags": ["teams"]
			}
		},
		"/api/teams/{teamId}/members": {
			"get": {
				"operationId": "teamMembersList",
				"parameters": [
					{
						"name": "teamId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/TeamMemberDto"
									}
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "List team members",
				"tags": ["teams"]
			},
			"post": {
				"operationId": "teamMembersAdd",
				"parameters": [
					{
						"name": "teamId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AddMemberDto"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TeamMemberDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Add a team member",
				"tags": ["teams"]
			}
		},
		"/api/teams/{teamId}/members/{userId}": {
			"patch": {
				"operationId": "teamMembersUpdateRole",
				"parameters": [
					{
						"name": "teamId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "userId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateMemberRoleDto"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/TeamMemberDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Update team member role",
				"tags": ["teams"]
			},
			"delete": {
				"operationId": "teamMembersRemove",
				"parameters": [
					{
						"name": "teamId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "userId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OkDto"
								}
							}
						}
					}
				},
				"security": [
					{
						"bearerAuth": []
					}
				],
				"summary": "Remove a team member",
				"tags": ["teams"]
			}
		},
...
	},
	"info": {
		"title": "kubwave API",
		"description": "",
		"version": "dev",
		"contact": {}
	},
...
	"servers": [],
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"type": "http"
			}
		},
		"schemas": {
...
			"TeamDto": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"name": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": ["owner", "member"]
					},
					"isDefault": {
						"type": "boolean"
					},
					"joinedAt": {
						"type": "string"
					}
				},
				"required": ["id", "name", "role", "isDefault", "joinedAt"]
			},
			"TeamStateDto": {
				"type": "object",
				"properties": {
					"teams": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/TeamDto"
						}
					},
					"activeTeamId": {
						"type": "string",
						"format": "uuid",
						"nullable": true
					}
				},
				"required": ["teams", "activeTeamId"]
			},
			"CreateTeamDto": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1,
						"maxLength": 100
					}
				},
				"required": ["name"]
			},
			"SetActiveTeamDto": {
				"type": "object",
				"properties": {
					"teamId": {
						"type": "string",
						"format": "uuid"
					}
				},
				"required": ["teamId"]
			},
			"ActiveTeamResponseDto": {
				"type": "object",
				"properties": {
					"activeTeamId": {
						"type": "string",
						"format": "uuid"
					},
					"team": {
						"$ref": "#/components/schemas/TeamDto"
					}
				},
				"required": ["activeTeamId", "team"]
			},
			"RenameTeamDto": {
				"type": "object",
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1,
						"maxLength": 100
					}
				},
				"required": ["name"]
			},
			"OkDto": {
				"type": "object",
				"properties": {
					"ok": {
						"type": "boolean"
					}
				},
				"required": ["ok"]
			},
			"TeamMemberDto": {
				"type": "object",
				"properties": {
					"userId": {
						"type": "string",
						"format": "uuid"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string",
						"format": "email"
					},
					"role": {
						"type": "string",
						"enum": ["owner", "member"]
					},
					"joinedAt": {
						"type": "string"
					}
				},
				"required": ["userId", "name", "email", "role", "joinedAt"]
			},
			"AddMemberDto": {
				"type": "object",
				"properties": {
					"email": {
						"type": "string",
						"format": "email"
					}
				},
				"required": ["email"]
			},
			"UpdateMemberRoleDto": {
				"type": "object",
				"properties": {
					"role": {
						"type": "string",
						"enum": ["owner", "member"]
					}
				},
				"required": ["role"]
			},
...
		}
	}
}